Refactor helpers to seperate files
